
\name{BhGLM-package}
\Rdversion{1.1.0}
\alias{BhGLM-package}
\alias{BhGLM}
\docType{package}
\title{
  Bayesian hierarchical GLMs and survival models, with applications to Genomics and Epidemiology
}

\description{
  This package provides functions for setting up and fitting various Bayesian hierarchical models (generalized linear models (GLMs), Cox survival models, 
  negative binomial models, and ordered logistic or probit regressions), 
  for numerically and graphically summarizing the fitted models, and for evaluating the predictive performance. 
  Several types of priors on the coefficients can be used: Student-t, double-exponential, and mixture double-exponential. The models are fitted by using fast algorithms for estimating posterior modes rather than MCMC. The methods can be used to analyze not only general data but also large-scale genomic data (i.e., detecting disease-associated genes or variants, predictive and prognostic modeling of diseases and traits).  
}

\details{
  \tabular{ll}{
  Package: \tab BhGLM\cr
  Type: \tab Package\cr
  Version: \tab 1.1.0\cr
  Date: \tab 2019-04-28\cr
  License: \tab MIT\cr
  LazyLoad: \tab yes\cr
  }
}

\author{
  Nengjun Yi, nyi@uab.edu
}


