% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake.cv}
\alias{fake.cv}
\title{I don't know what does this dataset does}
\format{A data frame that have data:
\describe{
...
}
 @source Idont know}
\usage{
fake.cv
}
\description{
I don't know
}
\keyword{datasets}
