% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smooth_terms.R
\name{plot_smooth_terms}
\alias{plot_smooth_terms}
\title{Plot splines according to mgcv pre-fit object}
\usage{
plot_smooth_terms(term_index, G, coefs, top = NULL)
}
\arguments{
\item{term_index}{numeric vector, the indices of spline terms what to plot}

\item{G}{The mgcv pre-fit object that contains smooth spline information}

\item{coefs}{Coefficients from the model}

\item{top}{textGrob, the title of the plot}
}
\value{

}
\description{
Plot splines according to mgcv pre-fit object
}
\examples{
N <- 1000
p <- 10

set.seed(1)

dat <- data.frame(sim_Bai_logistic(N, p)$dat) 
G <- gam(y ~ s(x1, bs = "ps", k=13)+s(x2, bs = "ps", k=15)+s(x3, bs = "ps", k=15)+s(x4, bs = "ps", k=15)+
s(x5, bs = "ps", k=15)+s(x6, bs = "ps", k=15)+s(x7, bs = "ps", k=15)+s(x8, bs = "ps", k=15)+
 s(x9, bs = "ps", k=15)+s(x10, bs = "ps", k=15), 
data=dat \%>\% data.frame(), family=binomial,drop.intercept = TRUE, fit = FALSE)

colnames(G$X) <- G$term.names
bgam_mdl <- bglm_spline(G$y~G$X-1, family = G$family, prior=mt(mean=0, s0=0.04, s1=1, df=Inf),
                        group = create_group(G))

plot_smooth_terms(1:4, G, coefficients(bgam_mdl),
                       top = textGrob("bgam with mgcv",gp=gpar(fontsize=20,font=3)) )
}
