% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_spline_terms.R
\name{test_spline_terms}
\alias{test_spline_terms}
\title{Use Wald test to test if the spline is necessary}
\usage{
test_spline_terms(G, coef, cov)
}
\arguments{
\item{G}{The mgcv pre-fit object that contains smooth spline information}

\item{cov}{}

\item{coefs}{Coefficients from the model}
}
\value{
A table of hypothesis testing results
}
\description{
Use Wald test to test if the spline is necessary
}
\examples{
#' N <- 1000
p <- 10

set.seed(1)

dat <- data.frame(sim_Bai_logistic(N, p)$dat) 
G <- gam(y ~ s(x1, bs = "ps", k=13)+s(x2, bs = "ps", k=15)+s(x3, bs = "ps", k=15)+s(x4, bs = "ps", k=15)+
s(x5, bs = "ps", k=15)+s(x6, bs = "ps", k=15)+s(x7, bs = "ps", k=15)+s(x8, bs = "ps", k=15)+
 s(x9, bs = "ps", k=15)+s(x10, bs = "ps", k=15), 
data=dat \%>\% data.frame(), family=binomial,drop.intercept = TRUE, fit = FALSE)

colnames(G$X) <- G$term.names
mdl <- bglm_spline(G$y~G$X-1, family = G$family, prior=mt(mean=0, s0=0.04, s1=1, df=Inf),
                        group = create_group(G))

test_groups(G, coefficients(mdl), vcov.bh(mdl))
}
